refactor(map): organize types of MapHistoryStore
