+++ frontend/src/features/undo_redo/index.ts
@@ -1,0 +1,5 @@
+import useMapState from './store/MapHistoryStore';
+
+export * from './store/state-types';
+export * from './store/action-types';
+export default useMapState;
