+++ frontend/src/features/undo_redo/store/state-types.ts
@@ -1,0 +1,52 @@
+/**
+ * A union type of the map layer's names.
+ */
+export type LayerName = 'plant';
+
+/**
+ * The state of a layer's object.
+ */
+export type ObjectState = {
+  index: LayerName;
+  id: string;
+  type: string;
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+  rotation: number;
+  scaleX: number;
+  scaleY: number;
+};
+
+/**
+ * The state of a map's layer.
+ */
+export type LayerState = {
+  index: LayerName;
+  visible: boolean;
+  /**
+   * The state of the objects on the layer.
+   */
+  objects: ObjectState[];
+};
+
+/**
+ * The state of the map.
+ */
+export type MapState = {
+  /**
+   * The state of the stage of the map.
+   */
+  stage: {
+    scale: number;
+    x: number;
+    y: number;
+    /**
+     * The state of the layers of the map.
+     */
+    layers: {
+      [key in LayerName]: LayerState;
+    };
+  };
+};
